refactor] Entities 수정
- Clean Architecture로 리팩토링 진행 중
	- Entities 폴더 추가하여 분리
